requirements.txt
flask
flask-cors
python-dotenv
groq
tavily-python
gunicorn
